Update Patrol/Undead looks, start adding 141-150, start adding mob variants for diff floors
Made Patrol/Undead labels display properly to the right of the target's name in a separate ImGui column
Started adding floors 141-150 (realized that some mobs that appear there appear again later in 190+ but should be treated more carefully hence why.......)
Started adding a way to tell which floor the player is on, if there are any repeat mobs (e.g. Deep Palace Knights) they can have multiple MobNotes that will be filtered based on which floor the player is currently on.