Add AggroTypeExtra and encounteredAgain
AggroTypeExtra is used to associate a FontAwesome icon to each aggro type, as well as a flavor text summarizing what the aggro type in question is.
encounteredAgain is used for mobs that will be present on multiple floors of palace (e.g. Deep Palace Knight), we're extending MobData to make it so we can know if the mob is ever encountered again on a later floor, and comparing that "later floor" to our current floor, we can offer different MobNotes as well as adjusted threat levels and more.